small tweak on the user-info panel
According to the official document of Bootstrap, tag <select> should be
avoided (via http://getbootstrap.com/components/#btn-groups). However I
do not know how to modify this issue due to the php code in it. The
consequence is that you have to click on the caret part of the genre
<select> dropdown menu, but not on the text part, so it means not the
whole button is clickable. And when you click on the caret, you can see
the layout changes.